requirements.txt
fastapi
pyjwt[crypto]
httpx
pyyaml
uvicorn
python-dotenv